[FIX] l10n_es_ticketbai_api: Don't put constraints on readonly fields
Moreover, such constraints are nonsense, as the fields are required on
the source model.

This fixes current warnings:

WARNING odoo odoo.models: method res.company._check_tbai_developer_id: @constrains parameter 'tbai_developer_id' is not writeable
WARNING odoo odoo.models: method res.company._check_tbai_license_key: @constrains parameter 'tbai_license_key' is not writeable
WARNING odoo odoo.models: method res.company._check_tbai_software_name: @constrains parameter 'tbai_software_name' is not writeable
